requirements.txt
asgiref==3.6.0
autopep8==1.6.0
black==23.1.0
click==8.1.3
Django==4.1.7
django-ckeditor==6.4.0
django-crontab==0.7.1
django-environ==0.8.1
django-extensions==3.2.1
django-filter==21.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-six==1.0.4
flake8==6.0.0
flake8-django==1.1.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.1.1
psycopg2==2.9.3
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
